requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.4
Django==6.0.2
django-cors-headers==4.7.0
djangorestframework==3.16.1
gunicorn==25.1.0
idna==3.11
packaging==26.0
PyJWT==2.11.0
python-arango==8.3.0
python-dotenv==1.2.2
requests==2.32.5
requests-toolbelt==1.0.0
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3